fix: created a component with test, reverted App.jsx
