+++ starters/solidjs-tanstack-tailwind/src/App.jsx
@@ -1,21 +1,7 @@
 import logo from './logo.svg';
 import styles from './App.module.css';
-import { createSignal } from 'solid-js';
 
 function App() {
-  const [todos, setTodos] = createSignal([]);
-
-  const toggleTodo = (value) => {
-    if(todos().includes(value)) {
-      const updateTodo = todos().filter(res => res !== value);
-      setTodo([...updateTodo])
-    } else if(todos().length > 0){
-      setTodos([...todos(), value])
-    } else {
-      setTodos([value])
-    }
-  }
-  const isChecked = (value) => todos().includes(value);
   return (
     <div class={styles.App}>
       <header class={styles.header}>
@@ -23,11 +9,10 @@ function App() {
         <p>
           Edit <code>src/App.jsx</code> and save to reload.
         </p>
-        <a class={styles.link} href="https://github.com/solidjs/solid" target="_blank" rel="noopener noreferrer" data-testid="learn-solidjs">
+        <a class={styles.link} href="https://github.com/solidjs/solid" target="_blank" rel="noopener noreferrer">
           Learn Solid
         </a>
       </header>
-      <input type="checkbox" checked={isChecked('check1')} onchange={[toggleTodo, 'check1']} />
     </div>
   );
 }
